feat: Home ダッシュボード化 + UI リッチモダン化 + 使い方ページ最新化
ログイン中のユーザーがホームを開いたとき「自分の今」が一目で分かるよう、
旧 /crests（殿堂）の中身をホームに集約。3 層階調 + soft elevation +
frosted glass + ページ遷移アニメで、操作 UX を保ったままモダンな見た目に。

## Home ダッシュボード化（ログイン時）

- 旧 /crests の中身（統計 / フィルタ / 進行中・達成カード / 紋章カタログ）を
  HallContent.tsx に切り出して Home に集約
- 上部に "── HOME ── / {ニックネーム}、本日も誓約の途上にあり" + StarDivider
- 下部に「新たな誓いを立てる」CTA
- 未ログインのホームは既存のヒーロー + 使い方 + 特徴をそのまま温存

## ルーティング統廃合

- /crests → / にリダイレクト（殿堂ページ廃止、CrestsGalleryPage.tsx 削除）
- /pacts → / にリダイレクト
- BottomTabs: 「殿堂」を廃止し「広場（/explore）🌳」に置換
  → 4 タブ: ホーム / 誓う / 広場 / 設定
- BottomTabs match で /pacts /crests も「ホーム」アクティブ判定
- PactDetailPage / Footer の /crests 参照を / または /explore に更新

## UI リッチモダン化

### 3 層階調
- 外側 parchment-bg の上に radial gradient で羊皮紙の奥行き
- 中間カードは parchment-card（最も明るい）に切替
  （HOW_IT_WORKS / 特徴 / ContractCard / FAQ / 機能一覧 全部）

### Soft elevation システム
- application.css に CSS 変数を定義:
  --shadow-card / --shadow-card-hover / --shadow-elevated
- TrophyCard / ActiveCard は onMouseEnter/Leave で shadow を切替
- ContractCard は --shadow-elevated に統一

### Frosted glass
- ヘッダー: parchment-bg/80 backdrop-blur-sm → parchment-bg/70 backdrop-blur-md
  + 微弱な金の bottom shadow
- BottomTabs: parchment-bg/95 → parchment-bg/85（透過感↑）

### ページ遷移アニメ
- Layout に motion.main を追加、location.pathname を key にして
  fade + 8px slide-up（duration 0.25s）
- prefers-reduced-motion を尊重（useReducedMotion）

## 使い方ページ最新化 + 言葉づかい平易化

### 内容更新
- Step 1: AI ジャンル指定 + 難易度判定の説明
- Step 2: 進捗草式の説明
- Step 3: 称号の自動付与、紋章コレクションは「ホーム画面」に
- FAQ: 達成判定の寛容モード、進行中契約 7 件上限、編集可能な属性、広場
  へのアクセス（計 8 問）
- 機能一覧: 殿堂→広場、進捗草式・称号を追加（9 機能）

### 言葉づかい平易化（エンジニア用語を排除）
- BottomTabs / /explore / is_public / POST / kept / broken / skipped 等の
  専門用語をすべて自然な日本語に置き換え
- 「× 遵守率 × 期間で総合スコア」→「守れた割合や期間を組み合わせた総合スコア」
- 「論理削除」→「破棄」、「不可逆」→「達成のままで残ります」など

## 検証

- bundle exec rspec: 338/338 examples passed
- bundle exec rubocop: clean
- npm run lint: clean
- npx tsc --noEmit: clean
