+++ app/frontend/pages/HomePage.tsx
@@ -2,6 +2,8 @@ import { Link, useNavigate } from "react-router-dom"
 import { useMutation, useQueryClient } from "@tanstack/react-query"
 import Layout from "../components/Layout"
 import Button from "../components/Button"
+import HallContent from "../components/HallContent"
+import StarDivider from "../components/StarDivider"
 import { useAuth } from "../hooks/useAuth"
 import { api, ApiError } from "../lib/api"
 import type { User } from "../types/user"
@@ -27,8 +29,16 @@ const HOW_IT_WORKS = [
   },
 ]
 
+/**
+ * ホームページ。
+ * - 未ログイン: マーケティングランディング（ヒーロー + 使い方 + 特徴）
+ * - ログイン中: ダッシュボード（ウェルカム + 統計 + フィルタ + 自分の契約 + 紋章カタログ）
+ *
+ * ログイン中ユーザーが「自分の今」を一目で把握できるように、旧 /crests
+ * （誓約の殿堂）のコンテンツをここに集約した。/crests は本ホームへリダイレクト。
+ */
 function HomePage() {
-  const { isAuthenticated, isLoading } = useAuth()
+  const { isAuthenticated, isLoading, user } = useAuth()
   const navigate = useNavigate()
   const queryClient = useQueryClient()
 
@@ -39,6 +49,30 @@ function HomePage() {
       navigate("/pacts/new/step1")
     },
   })
+
+  if (isLoading) {
+    return (
+      <Layout>
+        <p className="text-center text-ink/60 mt-12">読み込み中…</p>
+      </Layout>
+    )
+  }
+
+  if (isAuthenticated) {
+    return (
+      <Layout>
+        <div className="max-w-3xl mx-auto px-1">
+          <DashboardWelcome nickname={user?.nickname ?? "誓約者"} />
+          <HallContent />
+          <div className="text-center mt-8 mb-2">
+            <Button variant="primary" onClick={() => navigate("/pacts/new/step1")}>
+              新たな誓いを立てる
+            </Button>
+          </div>
+        </div>
+      </Layout>
+    )
+  }
 
   return (
     <Layout>
@@ -71,59 +105,46 @@ function HomePage() {
             自分との契約を、形にする場。
           </p>
 
-          {isLoading ? (
-            <p className="text-ink/60 text-sm">読み込み中…</p>
-          ) : isAuthenticated ? (
-            <div className="flex flex-col sm:flex-row gap-3 justify-center">
-              <Button variant="primary" onClick={() => navigate("/pacts/new/step1")}>
-                新たな誓約を結ぶ
+          <div className="flex flex-col items-center gap-4 w-full max-w-sm mx-auto">
+            <div className="flex flex-col sm:flex-row gap-2 w-full">
+              <Button variant="primary" fullWidth onClick={() => navigate("/signup")}>
+                新規登録
               </Button>
-              <Button variant="ghost" onClick={() => navigate("/crests")}>
-                誓約の殿堂へ
+              <Button variant="ghost" fullWidth onClick={() => navigate("/login")}>
+                ログイン
               </Button>
             </div>
-          ) : (
-            <div className="flex flex-col items-center gap-4 w-full max-w-sm mx-auto">
-              <div className="flex flex-col sm:flex-row gap-2 w-full">
-                <Button variant="primary" fullWidth onClick={() => navigate("/signup")}>
-                  新規登録
-                </Button>
-                <Button variant="ghost" fullWidth onClick={() => navigate("/login")}>
-                  ログイン
-                </Button>
-              </div>
 
-              <div className="flex items-center gap-2 w-full">
-                <span className="flex-1 border-t border-ink/20" />
-                <span className="text-xs text-ink/50">または</span>
-                <span className="flex-1 border-t border-ink/20" />
-              </div>
+            <div className="flex items-center gap-2 w-full">
+              <span className="flex-1 border-t border-ink/20" />
+              <span className="text-xs text-ink/50">または</span>
+              <span className="flex-1 border-t border-ink/20" />
+            </div>
 
-              <Button
-                variant="secondary"
-                fullWidth
-                onClick={() => guestMutation.mutate()}
-                disabled={guestMutation.isPending}
-              >
-                {guestMutation.isPending ? "支度中..." : "🎲 登録せずに 30 日お試し"}
-              </Button>
-              <p className="text-[11px] text-ink/50 -mt-2">
-                作った誓約は本登録時にそのまま引き継がれます
+            <Button
+              variant="secondary"
+              fullWidth
+              onClick={() => guestMutation.mutate()}
+              disabled={guestMutation.isPending}
+            >
+              {guestMutation.isPending ? "支度中..." : "🎲 登録せずに 30 日お試し"}
+            </Button>
+            <p className="text-[11px] text-ink/50 -mt-2">
+              作った誓約は本登録時にそのまま引き継がれます
+            </p>
+
+            {guestMutation.isError && (
+              <p className="text-xs text-seal">
+                ゲストの開始に失敗しました。再度お試しください。
               </p>
-
-              {guestMutation.isError && (
-                <p className="text-xs text-seal">
-                  ゲストの開始に失敗しました。再度お試しください。
-                </p>
-              )}
-            </div>
-          )}
+            )}
+          </div>
         </div>
       </section>
 
       {/* 使い方 3 ステップ（ファーストビューに収まる位置に） */}
       <section className="max-w-5xl mx-auto mb-12">
-        <div className="flex items-end justify-between mb-1">
+        <div className="flex items-end justify-between mb-4">
           <div>
             <p
               className="font-display font-semibold"
@@ -143,14 +164,12 @@ function HomePage() {
             くわしく →
           </Link>
         </div>
-        <div className="mb-4" />
-
 
         <div className="grid grid-cols-3 gap-2 sm:gap-4">
           {HOW_IT_WORKS.map((item) => (
             <div
               key={item.step}
-              className="relative p-3 sm:p-5 bg-parchment border border-gold/40 rounded-xl shadow-sm hover:shadow-md transition"
+              className="relative p-3 sm:p-5 bg-parchment-card border border-gold/40 rounded-xl shadow-sm hover:shadow-md hover:-translate-y-0.5 transition"
             >
               <div className="absolute -top-2 -left-2 w-7 h-7 sm:w-8 sm:h-8 flex items-center justify-center rounded-full bg-seal text-parchment font-serif font-bold text-xs sm:text-sm shadow">
                 {item.step}
@@ -207,7 +226,7 @@ function HomePage() {
           ].map((f) => (
             <div
               key={f.title}
-              className="p-5 bg-parchment/60 border border-gold/30 rounded-xl text-center hover:border-seal hover:shadow-sm transition"
+              className="p-5 bg-parchment-card border border-gold/30 rounded-xl text-center hover:border-seal hover:shadow-md hover:-translate-y-0.5 transition"
             >
               <div className="text-3xl mb-2">{f.emoji}</div>
               <h4 className="font-bold text-seal mb-1 text-sm">{f.title}</h4>
@@ -223,6 +242,44 @@ function HomePage() {
         </div>
       </section>
     </Layout>
+  )
+}
+
+/**
+ * ダッシュボード上部のウェルカムブロック。
+ * 「── HOME ──」キャプション + 「{ニックネーム}、本日も誓約の途上にあり」
+ */
+function DashboardWelcome({ nickname }: { nickname: string }) {
+  return (
+    <header className="text-center pt-4 pb-3">
+      <p
+        className="font-display font-semibold mb-2"
+        style={{
+          fontSize: 10,
+          letterSpacing: "0.55em",
+          color: "var(--color-gold-deep)",
+          paddingLeft: "0.55em",
+        }}
+        aria-hidden="true"
+      >
+        ── HOME ──
+      </p>
+      <h1
+        className="font-serif font-bold m-0"
+        style={{
+          fontSize: "clamp(20px, 5.5vw, 26px)",
+          letterSpacing: "0.1em",
+          color: "var(--color-ink)",
+          paddingLeft: "0.1em",
+          lineHeight: 1.4,
+        }}
+      >
+        <span className="text-seal">{nickname}</span>、本日も誓約の途上にあり
+      </h1>
+      <div className="mt-3">
+        <StarDivider />
+      </div>
+    </header>
   )
 }
